+++ CHANGELOG.markdown
@@ -1,4 +1,8 @@
-3.7.1.1
+3.7.1.2 [maintenance release]
+-------
+* Made the doctest test suite hide all but the exact versions of packages used to build this package to avoid problems with complicated user environments.
+
+3.7.1.1 [maintenance release]
 -------
 * Removed tests that will (likely) fail in the presence of `hashable` 1.2